Move filter into its own controller and add extra tests
